Modify test namelist
After a discussion within the DLR-core group this new namelist design was agreed on. However, it is expected that the namelist design will still change and adapt in future.

The preprocessing tasks can be defined outside the diagnostic blocks and be referenced at the declaration of the variables inside the single diagnostic blocks. Furthermore, the diagnostic blocks contain the "tags"-Element.