Moved time-average operator after the multimodel statistics preprocessor:
Discussion is here:
https://github.com/ESMValGroup/ESMValTool/pull/438#discussion_r199066818
"It's conceptually not correct to apply time/area statistics before the multi-model. This is because, in general, the multi-model of some stats is not equal to the stats of the multi-model.
Since the multi-model-mean/median is treated like another dataset, it's important that the preprocessing is applied consistently to all of them."