Explain how to download a single file from github
First time users that use conda to install esmvaltool might have trouble getting to the config-user file. The docs point to the config-user-example.yml on github, but for those new users it may not be clear how to download a single file from there. This should make it a little bit more accessible.