Add option to use variable change as target
For the perfect model test a target (such as temperature in the future)
needs to be set. This update allows to also set temperature change
(to a historical reference period) as target. All needed variables
are already being calculated (for the temperture change maps - which
show exactly that). I've moved the calculation to
calculate_weights_climwip, and added a small routine calculating the
change to calibrate_sigmas.